Move activities
Put activities in their own directory.
