+++ app/src/main/java/com/gophillygo/app/activities/HomeActivity.java
@@ -1,4 +1,4 @@
-package com.gophillygo.app;
+package com.gophillygo.app.activities;
 
 import android.arch.lifecycle.ViewModelProviders;
 import android.content.Intent;
@@ -15,6 +15,10 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.GridView;
 
+import com.gophillygo.app.CarouselViewListener;
+import com.gophillygo.app.R;
+import com.gophillygo.app.activities.EventsListActivity;
+import com.gophillygo.app.activities.PlacesListActivity;
 import com.gophillygo.app.adapters.PlaceCategoryGridAdapter;
 import com.gophillygo.app.data.DestinationViewModel;
 import com.gophillygo.app.data.models.Destination;
