+++ app/src/main/res/layout/activity_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.gophillygo.app.HomeActivity">
+    tools:context="com.gophillygo.app.activities.HomeActivity">
 
     <com.synnapps.carouselview.CarouselView
         android:id="@+id/home_carousel"
