+++ app/src/main/res/layout/activity_places_list.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.gophillygo.app.PlacesListActivity">
+    tools:context="com.gophillygo.app.activities.PlacesListActivity">
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
